+++ draft/EinschreibenWaehlerliste.java
@@ -17,9 +17,9 @@ public class EinschreibenWaehlerliste {
     public void einschreibungen(){
         for(int i=0; i<buergerListe.size();i++){
             BigInteger[] loc = new BigInteger[3];
-            loc[0] = buergerListe.get(i).getE();
-            loc[1] = buergerListe.get(i).getN();
-            loc[2] = buergerListe.get(i).getZ();
+            loc[0] = buergerListe.get(i).getE().modPow(mix.getPublicE(), mix.getPublicN());
+            loc[1] = buergerListe.get(i).getN().modPow(mix.getPublicE(), mix.getPublicN());
+            loc[2] = buergerListe.get(i).getZ().modPow(mix.getPublicE(), mix.getPublicN());
             mix.empfange(loc);            
         }
         System.out.println("mix beschrieben");
